1. move docker handler to common lib 2. add java project class to common lib with should be shared 3. fix typo: 'UNKNOW'
