+++ azure-tools-common/src/main/java/com/microsoft/azure/maven/function/Constants.java
@@ -10,9 +10,9 @@ public class Constants {
     public static final String APP_SETTING_WEBSITE_RUN_FROM_PACKAGE = "WEBSITE_RUN_FROM_PACKAGE";
     public static final String APP_SETTING_MACHINEKEY_DECRYPTION_KEY = "MACHINEKEY_DecryptionKey";
     public static final String APP_SETTING_WEBSITES_ENABLE_APP_SERVICE_STORAGE = "WEBSITES_ENABLE_APP_SERVICE_STORAGE";
-    public static final String APP_SETTING_WEBSITES_ENABLE_APP_SERVICE_STORAGE_VALUE = "false";
+    public static final String APP_SETTING_DISABLE_WEBSITES_APP_SERVICE_STORAGE = "false";
     public static final String APP_SETTING_FUNCTION_APP_EDIT_MODE = "FUNCTION_APP_EDIT_MODE";
-    public static final String APP_SETTING_FUNCTION_APP_EDIT_MODE_VALUE = "readOnly";
+    public static final String APP_SETTING_FUNCTION_APP_EDIT_MODE_READONLY = "readOnly";
     public static final String ZIP_EXT = ".zip";
     public static final String LOCAL_SETTINGS_FILE = "local.settings.json";
     public static final String INTERNAL_STORAGE_KEY = "AzureWebJobsStorage";
