+++ azure-tools-common/src/main/java/com/microsoft/azure/maven/function/handlers/artifact/DockerArtifactHandler.java
